an error is encountered.

+++ systems/lorenz.py
@@ -1,8 +1,8 @@
 import numpy as np
-def lorenz(state,sigma=10,rho=28,beta=8/3):
-    x,y,z=state
-    dx=sigma*(y-x)
-    dy=x*(rho-z)-y
-    dz=x*y-beta*z
-    
-    return np.array([dx,dy,dz])
+
+def lorenz(state, sigma=10.0, rho=28.0, beta=8.0/3.0):
+    x, y, z = state
+    dx = sigma * (y - x)
+    dy = x * (rho - z) - y
+    dz = x * y - beta * z
+    return np.array([dx, dy, dz], dtype=np.float64)
